pyproject.toml
[project]
name = "tlaevolve"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = ["openevolve", "python-dotenv"]
